Bumped to version 0.5 -- ATTENTION -- BREAKING CHANGES
Removed the untyped structure tag `scalar`.
Replaced it with five new tags providing type information:

  - `bool`
  - `float`
  - `int`
  - `null`
  - `string`

When reading YAML data types are now determined as per the YAML Core
schema rules. And the assumption that quoted values are always
strings, and none of the other types.

Read values are now normalized, as per their type. For example, the
internally seen value of an integer is always in decimal form,
regardless of if it was octal or hex in the input.

On the writing side the new structural tags now exist as pre-defined
type converters. The `scalar` converter is still present, and aliases
to `string`. In the previous version `string` aliased to `scalar` as
the only scalar type.

Most styling of scalar values is now automatically determined by their
type, and, in case of strings, by the value as well.

Further changes, non-breaking:

  - Added a testsuite.
  - Updated the documentation to match the changes above.
  - Regenerated the dependent files.